requirements.txt
Django==1.9.7
Pillow==3.3.0
cloudinary==1.4.0
python-magic==0.4.12
requests==2.10.0
